add matcher engine

+++ edge.go
@@ -13,9 +13,9 @@ import (
 	Polyline - geometry of edge, pointer to s2.Polyline (wrapper)
 */
 type Edge struct {
-	ID     int
-	Source int
-	Target int
+	ID     int64
+	Source int64
+	Target int64
 	Weight float64
 	*s2.Polyline
 }
